feat: updated logger for database
