+++ server/internal/db/main.go
@@ -3,26 +3,22 @@ package db
 import (
 	"github.com/jmoiron/sqlx"
 	_ "github.com/mattn/go-sqlite3"
+	"github.com/mr55p-dev/pagemail/internal/logging"
 )
 
-var PageEventMap map[string]EventOutput[Page]
-var log logging.Log
-
-func init() {
-	log = logging.GetLogger("db")
-	PageEventMap = make(map[string]EventOutput[Page])
-}
-
 type Client struct {
-	db *sqlx.DB
+	log logging.Logger
+	db  *sqlx.DB
 }
 
 func (c *Client) Close() {
 	c.db.Close()
 }
 
-func NewClient(path string) *Client {
+func NewClient(path string, log logging.Logger) *Client {
 	conn := sqlx.MustOpen("sqlite3", path)
-
-	return &Client{conn}
+	return &Client{
+		log: log,
+		db:  conn,
+	}
 }
